requirements.txt
requests==2.31.0
python-dotenv==1.0.0
atproto==0.0.36
folium==0.14.0
geopandas==0.13.2
contextily==1.3.0
matplotlib==3.7.2
shapely==2.0.1
pandas==2.1.0
Pillow==10.1.0
SQLAlchemy==2.0.20
PyYAML==6.0.1
